Extract thread classification module
